% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Adherence_dashboard.R
\name{adherence_dashboard}
\alias{adherence_dashboard}
\title{Dashboard of chronic medication adherence}
\usage{
adherence_dashboard(calculated_adherence_w_info)
}
\arguments{
\item{calculated_adherence_w_info}{This is supposed to be previously calculated CMAs of chronic medication consumers and their medication included with additional information of consumers (age, age group, gender) and additional information of medication (ATC code)}
}
\value{
returns a Shiny app with the possibilty to observe changes in medication adherence between genders, age groups and specific medication groups. The app has three main plots. First one shows medication adherence changes through time between male and female consumers. Second one shows total average medication adherence between different medication groups. Third one shows chenges in medication adherence between age groups. The app makes it possible to filter out specific genders and ages, as well as  filter out medication based on five different ATC levels (you may want to include only medication with ATC I level "A" or get more detailed with ATC II-V level for example A06AH04 - Naloxone).
}
\description{
Dashboard of chronic medication adherence
}
\examples{
adherence_dashboard(calculated_adherence_with_additional_info_added_with_add_info_to_adherence_function)
}
