% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_dashboard_summarized.R
\name{adherence_dashboard_summarized}
\alias{adherence_dashboard_summarized}
\title{Interactive adherence dashboard with easy analysis genereated with summarized function adherence_summary.}
\usage{
adherence_dashboard_summarized(summarized_adherence)
}
\arguments{
\item{summarized_adherence}{Summarized table generated with function adherence_summary}
}
\value{
returns Shiny dashboard with interactive possibilities to observe adherence in different groupings. Using adherence summary and dashboard for said table is more efficient and optimized way to analyse populations adherence
}
\description{
Interactive adherence dashboard with easy analysis genereated with summarized function adherence_summary.
}
\examples{
adherence_dashboard_summarized(summarized_table)
}
