% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherence_summary.R
\name{adherence_summary}
\alias{adherence_summary}
\title{Summarize the adherence table for convenient analysis and optimized dashboard generation}
\usage{
adherence_summary(df_w_info)
}
\arguments{
\item{df_w_info}{This should be the calculated adherence table (done with function 'adherence_function') with the additional information about the patients (age, gender) as well as information about the purchased medication (ATC code for easy filtering). The information should be added to the adherence table with the funcyion 'add_info_to_adherence'.}
}
\value{
Function returns a summarized table with all possible combinations of gender, age-groups, medications and time window groupings. Summarized information includes information about average, median, count, maxiumum, minimum and quantiles of the CMAs.
}
\description{
Summarize the adherence table for convenient analysis and optimized dashboard generation
}
\examples{
summarized_table<-adherence_summary(CMA6_w_information)
}
