% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherenceCalculations.R
\name{adherenceCalculations}
\alias{adherenceCalculations}
\title{Calculating simple CMAs for the duration on observation period}
\usage{
adherenceCalculations(
  cdm = NULL,
  data = NULL,
  cma = c("CMA1", "CMA2", "CMA3", "CMA4", "CMA5", "CMA6", "CMA7", "CMA8", "CMA9"),
  name = NULL,
  medicationGroup = NULL,
  ...
)
}
\arguments{
\item{cdm}{a cdm_reference object with 'name' table}

\item{data}{instead of cdm provide in-memory data.}

\item{cma}{By default the function calculates adherence with all different CMA algorithms (CMA1-CMA9).}

\item{name}{name of the table used for adherence calculations}

\item{medicationGroup}{CMA is calculated for each defined group. Medication group is a named list with a name and drug_concept_id values. For creating this list it is highly recommended to use CodelistGenerator \link[CodelistGenerator]{"https://darwin-eu.github.io/CodelistGenerator/articles/a04_GenerateVocabularyBasedCodelist.html"}}

\item{...}{input values for adhereR CMA function call}
}
\value{
table with CMA results
}
\description{
Calculating simple CMAs for the duration on observation period
}
\examples{
\dontrun{
connection <- DBI::dbConnect(duckdb::duckdb(dbdir = CDMConnector::eunomiaDir()))

cdm <- CDMConnector::cdmFromCon(
  con = connection,
  cdmSchema = "main",
  writeSchema = "main",
  cdmName = "my_duckdb_database"
)

cdm <- generateChronicDrugExposure(
  cdm = cdm,
  ingredients = c("acetaminophen", "metformin", "simvastatin"),
  name = "test_table",
  overwrite = TRUE
)
adherenceCalculations(
  cdm = cdm,
  ingredient = c("acetaminophen"),
  name = "test_table",
  cma = "CMA1"
)
}
}
