% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherenceCalculations.R
\name{adherenceCalculationsLoader}
\alias{adherenceCalculationsLoader}
\title{For using adherence calculations function with large datasets.}
\usage{
adherenceCalculationsLoader(
  reference,
  cdm,
  resultsTableName = "adherenceFromOMOP_results",
  cma = c("CMA1", "CMA2", "CMA3", "CMA4", "CMA5", "CMA6", "CMA7", "CMA8", "CMA9"),
  batchSize = 1000,
  ...
)
}
\arguments{
\item{reference}{reference to table}

\item{cdm}{a cdm_reference object with 'name' table}

\item{resultsTableName}{name of the table where results are stores}

\item{cma}{By default the function calculates adherence with all different CMA algorithms (CMA1-CMA9).}

\item{batchSize}{default 1000.}

\item{...}{input values for adherenceCalculations}
}
\description{
Calculates adherence in batches
}
\examples{
\dontrun{
cdm <- mockAdherenceFromOMOP()

chronicDrugExposure <- generateChronicDrugExposure(cdm = cdm, name = "chronic_drug_exposure_table", overwrite = T)

CMA_values <- adherenceCalculationsLoader(reference = chronicDrugExposure, cdm = cdm, cma = "CMA5")
}

}
