% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{adherenceDataDoc}
\alias{adherenceDataDoc}
\title{Documentation helper for adherenceData parameter}
\arguments{
\item{adherenceData}{Database table reference or data frame containing adherence
calculation results, typically generated by [calculateAdherence()] or
[calculateAdherenceSlidingWindow()].}
}
\description{
Documentation helper for adherenceData parameter
}
\keyword{internal}
