% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherenceCalculations.R
\name{adherenceSlidingWindowLoader}
\alias{adherenceSlidingWindowLoader}
\title{For using adherence sliding window function with large datasets}
\usage{
adherenceSlidingWindowLoader(
  reference,
  cdm,
  resultsTableName = "adherenceFromOMOP_results",
  cma = c("CMA1", "CMA2", "CMA3", "CMA4", "CMA5", "CMA6", "CMA7", "CMA8", "CMA9"),
  medicationGroup = NULL,
  batchSize = 10000,
  ...
)
}
\arguments{
\item{reference}{reference to table}

\item{cdm}{a cdm_reference object}

\item{resultsTableName}{name of the table where results are stores}

\item{cma}{By default the function calculates adherence with all different CMA algorithms (CMA1-CMA9).}

\item{medicationGroup}{CMA is calculated for each defined group. Medication group is a named list with a name and drug_concept_id values. For creating this list it is highly recommended to use CodelistGenerator \link[CodelistGenerator]{"https://darwin-eu.github.io/CodelistGenerator/articles/a04_GenerateVocabularyBasedCodelist.html"}}

\item{batchSize}{default 1000.}

\item{...}{input values for CMA_sliding_window}
}
\description{
Calculates adherence in batches
}
