% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cdmDoc}
\alias{cdmDoc}
\title{Documentation helper for cdm parameter}
\arguments{
\item{cdm}{(`cdm_reference`) A CDM reference object created with CDMConnector.}
}
\description{
Documentation helper for cdm parameter
}
\keyword{internal}
