% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronic_drug_exposure_query.R
\name{chronic_drug_exposure}
\alias{chronic_drug_exposure}
\title{Get specific drug exposure entries for adherence calculations}
\usage{
chronic_drug_exposure(connection, schema, ingredients = ing_4_sql)
}
\arguments{
\item{connection}{Database connection based on individual connection details (ordinarily as conn-variable)}

\item{schema}{String of OMOP schema which includes the necessary drug_exposure table from which adherence can be calculated.}

\item{ingredients}{Specifies medication under observation. Uses RxNorm Ingredient concept code as acceptable parameter (Function converts drug_exposure drug_concept_id into specific ingredients which he medication consists of). By default the parameter uses ing_4_sql, which are so called mapped chronic medication ingredients but can be adjusted for preference}
}
\value{
Dataframe of drug_exposure entries which consists of necessary columns for adherence calculation
}
\description{
Get specific drug exposure entries for adherence calculations
}
\examples{
chronic_drug_exposure(connection=conn, schema='ohdsi_cdm_202207')
chronic_drug_exposure(connection=conn, schema='ohdsi', ingredients=c('1502905', '1580747,1503297'))
chronic_drug_exposure(connection=conn, schema='ohdsi_cdm_202208', ingredients = '1322184')
}
