% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chronic_drug_exposure_query.R
\name{chronic_drug_exposure_query}
\alias{chronic_drug_exposure_query}
\title{Query to get chronic drug exposure drug events from OMOP drug_exposure table}
\usage{
chronic_drug_exposure_query(schema)
}
\arguments{
\item{schema}{(string of OMOP schema, with chronic drug exposure)}
}
\value{
query for querysql
}
\description{
Query to get chronic drug exposure drug events from OMOP drug_exposure table
}
\examples{
chronic_drug_exposure_query('ohdsi_cdm_202207')
}
