% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{cmaDoc}
\alias{cmaDoc}
\title{Documentation helper for cma parameter}
\arguments{
\item{cma}{(`character`) Vector of CMA types to calculate. Options:
`"CMA1"` through `"CMA9"`. See [AdhereR documentation](https://www.adherer.eu/)
for details on each CMA type.}
}
\description{
Documentation helper for cma parameter
}
\keyword{internal}
