% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{conceptSetDoc}
\alias{conceptSetDoc}
\title{Documentation helper for conceptSet parameter}
\arguments{
\item{conceptSet}{(`list`, `codelist`, or `NULL`) Named list of drug concept IDs.
Recommended to use [CodelistGenerator::getDrugIngredientCodes()].
If `NULL`, all drugs in the drug_exposure table are included.}
}
\description{
Documentation helper for conceptSet parameter
}
\keyword{internal}
