% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{indexDateDoc}
\alias{indexDateDoc}
\title{Documentation helper for indexDate parameter}
\arguments{
\item{indexDate}{(`character(1)`) Name of the column to use as the index date
for patient-level calculations. Default: `"drug_exposure_start_date"`.}
}
\description{
Documentation helper for indexDate parameter
}
\keyword{internal}
