% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{medicationGroupDoc}
\alias{medicationGroupDoc}
\title{Documentation helper for medicationGroup parameter}
\arguments{
\item{medicationGroup}{(`list` or `NULL`) Named list where each element contains
drug_concept_id values for a medication group. CMA is calculated separately for
each group. If `NULL`, all medications are treated as one group.}
}
\description{
Documentation helper for medicationGroup parameter
}
\keyword{internal}
