% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherenceCalculations.R
\name{medicationGrouping}
\alias{medicationGrouping}
\title{Function for grouping medications}
\usage{
medicationGrouping(cdm, ingredientGroups = NULL)
}
\arguments{
\item{cdm}{(`cdm_reference`) A CDM reference object created with CDMConnector.}

\item{ingredientGroups}{previously determined grouping, format as list(name = c(ingredient, ingredient)). Ingredient can be marked as concept ID or ingredient name.}
}
\value{
codelist generator output with the structure of ingredientGroups
}
\description{
Uses CodelistGenerator to get drug ingredient codes
}
\examples{
\dontrun{

 mock_cdm <- CodelistGenerator::mockVocabRef()
groups <- medicationGrouping(mock_cdm, ingredientGroups=list(group1 = c("Adalimumab")))
}
}
