% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentationHelpers.R
\name{nameDoc}
\alias{nameDoc}
\title{Documentation helper for name parameter}
\arguments{
\item{name}{(`character(1)`) Name for the output table in the database.}
}
\description{
Documentation helper for name parameter
}
\keyword{internal}
