% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{addBMI}
\alias{addBMI}
\title{adds bmi values to CMA values}
\usage{
addBMI(adherenceData, cdm)
}
\arguments{
\item{adherenceData}{Database table reference or data frame containing adherence
calculation results, typically generated by [calculateAdherence()] or
[calculateAdherenceSlidingWindow()].}

\item{cdm}{(`cdm_reference`) A CDM reference object created with CDMConnector.}
}
\value{
Adherence data with BMI category information added.
}
\description{
adds bmi values to CMA values
}
