% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{addBMI}
\alias{addBMI}
\title{adds bmi values to CMA values}
\usage{
addBMI(cma_table, cdm)
}
\arguments{
\item{cma_table}{reference to a table in database or one in memory that contains cma values}

\item{cdm}{a cdm_reference object}
}
\description{
adds bmi values to CMA values
}
