% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{addCohorts}
\alias{addCohorts}
\title{Adds cohort columns to cma data}
\usage{
addCohorts(adherenceData, cohort, cohortIdMappingToNames = NULL)
}
\arguments{
\item{adherenceData}{Database table reference or data frame containing adherence
calculation results, typically generated by [calculateAdherence()] or
[calculateAdherenceSlidingWindow()].}

\item{cohort}{Cohort table reference or in-memory table with columns:
cohort_definition_id, subject_id, cohort_start_date, cohort_end_date.}

\item{cohortIdMappingToNames}{(`list` or `NULL`)
Optional mapping of cohort IDs to names.}
}
\value{
Reference to table with adherence and cohort data.
}
\description{
Adds cohort columns to cma data
}
\examples{
\dontrun{
cdm <- mockDrugExposure()

chronicDrugExposure <- generateChronicDrugExposure(cdm = cdm, name = "chronic_drug_exposure_table", overwrite = T)

CMA_values <- calculateAdherenceBatched(drugExposure = chronicDrugExposure, cdm = cdm, cma = "CMA5")

observ_dates <- cdm$observation_period \%>\%
  dplyr::filter(person_id == 1) \%>\%
  dplyr::select(observation_period_start_date, observation_period_end_date) \%>\%
  dplyr::collect()

cohortTable <- dplyr::tibble(
  cohort_definition_id = 1,
  subject_id = 1,
  cohort_start_date = observ_dates$observation_period_start_date,
  cohort_end_date = observ_dates$observation_period_end_date
)

withCohort <- addCohorts(CMA_values, cohortTable)
}
}
