% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{addCohorts}
\alias{addCohorts}
\title{Adds cohort columns to cma data}
\usage{
addCohorts(cmaTable, cohortTable, cohortIdMappingToNames = NULL)
}
\arguments{
\item{cmaTable}{reference to cma table}

\item{cohortTable}{cohort table that is used for adding features. reference or in memory table with columns cohort_definition_id subject_id cohort_start_date cohort_end_date}
}
\value{
reference to table with cma and cohort data
}
\description{
Adds cohort columns to cma data
}
