% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherenceCalculations.R
\name{addPatientInformation}
\alias{addPatientInformation}
\title{Adds patient information generated by generate chronic drug exposure function. User does not call this function directly.}
\usage{
addPatientInformation(data, cma_values)
}
\arguments{
\item{data}{input used for AdhereR functions, generated with generateChronicDrugExposure}

\item{cma_values}{output of AdhereR functions}
}
\value{
cma values with patient information
}
\description{
Adds patient information generated by generate chronic drug exposure function. User does not call this function directly.
}
