% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addFeatures.R
\name{countPeople}
\alias{countPeople}
\title{Counts people, returns an overview of subjects}
\usage{
countPeople(
  cdm = NULL,
  cmaTable = NULL,
  generatedTable = NULL,
  cohortSchema = NULL,
  cohortId = NULL,
  n = 1
)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}

\item{cmaTable}{reference or in memory dataframe that contains CMA values}

\item{generatedTable}{reference or in memory dataframe that contains generateChronicDrug exposure output}

\item{cohortSchema}{string value of a schema in database}

\item{cohortId}{vector with cohort ids to count people included}

\item{n}{minimum number of rows per year for a person to be included in 'Yearly drug exposure record count' table}
}
\value{
list
}
\description{
Counts people, returns an overview of subjects
}
\examples{
\dontrun{
cdm <- mockAdherenceFromOMOP()
chronicDrugExposure <- generateChronicDrugExposure(cdm = cdm, name = "chronic_drug_exposure_table", overwrite = T)
CMA_values <- adherenceCalculationsLoader(reference = chronicDrugExposure, cdm = cdm, cma = "CMA5")
counts <- countPeople(cdm = cdm, cmaTable = CMA_values, generatedTable = chronicDrugExposure)
}
}
