% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateChronicDrugExposure.R
\name{generateChronicDrugExposure}
\alias{generateChronicDrugExposure}
\title{Get drug exposure entries for adherence calculations}
\usage{
generateChronicDrugExposure(
  cdm,
  name = "adherenceFromOMOP",
  indexDateForPatientCalculation = "drug_exposure_start_date",
  overwrite = FALSE,
  subjects = NULL,
  drugConceptIdList = NULL,
  cohortSchema = NULL,
  cohortId = NULL,
  cohortTableName = "cohort",
  ...
)
}
\arguments{
\item{cdm}{A cdm reference via CDMConnector.}

\item{name}{Name of the new cohort table, it must be a length 1 character vector.}

\item{indexDateForPatientCalculation}{Value for indexDate in addDeathDays function from PatientProfiles.}

\item{overwrite}{If table with provided name exists this value chooses the action.}

\item{subjects}{Possibility to choose a specific set of subjects.}

\item{drugConceptIdList}{list with drug concept ids that determine if drug exposure records are included or not}

\item{cohortSchema}{Possible include only certain people that exist in a specific defined cohort between a time period. Parameter cohort schema must be a string value of the database schema with the defined cohort.}

\item{cohortId}{List of numeric values of the defined cohorts that haveS included person ids as well as start and end date of the cohort.}

\item{cohortTableName}{usually named "cohort"}

\item{...}{Arguments to be passed to CodelistGenerator \link[CodelistGenerator]{getDrugIngredientCodes} function.}
}
\value{
reference to table
}
\description{
Get drug exposure entries for adherence calculations
}
\examples{
\dontrun{
cdm <- mockAdherenceFromOMOP()

chronicDrugExposure <- generateChronicDrugExposure(
  cdm = cdm,
  drugConceptIdList = concepts_ids,
  name = "chronic_drug_exposure_table",
  overwrite = T
)
}
}
