% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adherenceCalculations.R
\name{loadData}
\alias{loadData}
\title{Loads data, performs basic checks and mutates columns. User does not call this function directly.}
\usage{
loadData(cdm, name, data)
}
\arguments{
\item{cdm}{a cdm_reference object}

\item{name}{name of the table in database that contains output of generateChronicDrugExposure}

\item{data}{in-memory dataframe}
}
\value{
verified dataframe in a suitable format
}
\description{
Loads data, performs basic checks and mutates columns. User does not call this function directly.
}
