% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mockAdherenceFromOMOP.R
\name{mockAdherenceFromOMOP}
\alias{mockAdherenceFromOMOP}
\title{Mock database for testing package.}
\usage{
mockAdherenceFromOMOP(nPerson = 5, recordPerson = 3, dbms = duckdb::duckdb())
}
\arguments{
\item{nPerson}{numeric value for omock::mockPerson \link[omock]{mockPerson} function.}

\item{recordPerson}{numeric value for omock::mockDrugExposure \link[omock]{mockDrugExposure} function.}

\item{dbms}{currently accepts only duckdb}
}
\value{
CDM reference with necessary tables
}
\description{
Mock database for testing package.
}
\examples{
\dontrun{
cdm <- mockAdherenceFromOMOP()
}
}
